remove one warning: luatexini.c:1187: warning: format ‘%ld’ expects type ‘long int’, but argument 7 has type ‘pool_pointer’
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@1462 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/texmfmp.h
@@ -283,9 +283,9 @@ extern void paintrow (/*screenrow, pixelcolor, transspec, screencol*/);
     undumpthings (base, len);                                           \
     for (i = 0; i < (len); i++) {                                       \
       if ((&(base))[i] < (low) || (&(base))[i] > (high)) {              \
-        FATAL5 ("Item %u (=%ld) of .fmt array at %lx <%ld or >%ld",     \
+        FATAL5 ("Item %u (=%lu) of .fmt array at %lx <%lu or >%lu",     \
                 i, (unsigned long) (&(base))[i], (unsigned long) &(base),\
-                (unsigned long) low, (integer) high);                   \
+                (unsigned long) low, (unsigned long)((integer) high));                   \
       }                                                                 \
     }									\
   } while (0)
@@ -488,9 +488,9 @@ extern void zwclose P1H(FILE *);
     undump_things (base, len);                                           \
     for (i = 0; i < (len); i++) {                                       \
       if ((&(base))[i] < (low) || (&(base))[i] > (high)) {              \
-        FATAL5 ("Item %u (=%ld) of .fmt array at %lx <%ld or >%ld",     \
+        FATAL5 ("Item %u (=%lu) of .fmt array at %lx <%lu or >%lu",     \
                 i, (unsigned long) (&(base))[i], (unsigned long) &(base),     \
-                (unsigned long) low, (integer) high);                         \
+                (unsigned long) low, (unsigned long) high);             \
       }                                                                 \
     }																	\
   } while (0)
